feat(atom): connect RAM output to InstrReg input
Right now, the only way to get an instruction is to get it directly from
RAM, so I am hooking it up directly.

I wonder if it would ever be interesting to get an instruction from
somewhere else. Could you do something interesting by constructing an
instruction in a register and then executing it?