feat(atom): setup binary file import
Digital supports reading a binary file into RAM at startup. I will be
using these in order to load programs.

I will not be including the file path in the configuration of
hardware.dig, because it will differ from system to system. If you want
to run it, you will need to point the path to the .bin file on your
system. If I figure out how to have a relative path or contribute that
back to digital, I will use that because right now it is annoying to
share the program.

This basic program just sets AB in the lower bits of register 0.