feat(atom): add flags to ALU
The ALU has two flags. The first reports if the operation resulted in a
value that is negative (assuming two's complement). The second reports
if the value is zero.

If we perform subtraction a - b, then we can use these flags to compare
a and b. If the result is zero, a is equal to b. If the result is
negative, a is smaller than b.

This system is not perfect, and we need information about carry and
overflow values to make it more complete, which we will add soon.

The flags are saved when we execute a register operation.