refactor(atom): remove JLD control line
The JLD control line sent data from the jump calculation to the data
bus. However, it always wants to be high when the Jmp control line is
high. Right now, there is no use for sending a jump calculation to the
bus, so I just re-routed it to always go straight to the PC.