feat(atom): add memory read and write instructions
The memory read and write instructions are able to make pretty heavy use
of existing circuitry. The instructions follow the same basic shape of
register operations, and so are able to share how they send the register
identifiers to the register file.

For the immediate value, I needed to make a change that extends the 4
bit immediate value to the 8 bits expected for an immediate. It is a bit
weird, because later that gets extended again for the purpose of adding
it to the 16 bit register value, but it keeps the control unit pretty
simple.

Memory addressing logic had to be updated because I need to switch
between the PC and the computed memory address. For now, I switch
between them based upon the execution phase. This is okay because we
only will ever send the value to be bus during execution, and will never
update the instruction register during execution.

There also were to new control lines, one to send the data from memory
to the data bus as it cannot always be sent, and one to send the value
of the target register to the data bus when we want to write it.