feat(atom): add call instruction
The call instruction allows you to jump to a procedure and, after the
ret instruction is implemented later, return from it.

This requires the addition of a stack pointer register, which points to
memory where the stack can be written. The stack is offset by 0x8000 so
that it doesn't overlap with program memory for now, but it will
eventually reach and overwrite it.