feat(atom): improve observability
As the complexity of Atom increases, I need to be able to see and keep
track of the state of the computer over time. In Digital, that means
adding probes. You can always view the state of the probes. More
interestingly, you can also see a graph or table of the values over time
which can be useful when finding issues.

Most of the probes are just values within the computer. I did add an
interesting probe that counts clock cycles. There are a couple uses I
can see for this:
- Counting how many cycles a program took
- Stepping through a program back to the same point