docs(atom): add fibonacci with recursive optimization
Because the fibonacci procedure only calls itself, I have control over all the
registers. That allows me to make this optimization by setting up a register
only on the first call. All subsequent calls jump to the instruction after that,
saving a few instructions. Since the procedure is called many times, this is a
reasonable savings.