feat(atom): add peripheral instructions in hardware
I use EEPROM as storage because you can read and write to it, and it gets saved
even when the computer is off. It is actually saved into the .dig file, so it
can be preserved here. This is only true if you are executing the circuit, so
the EEPROM must be directly included in hardware.dig.