feat(atom): add simple print command
I have gotten pretty fast at assembling the code, and I was able to
assemble the full editor program in about 5 minutes. However, the
process is still error prone and it can be difficult to tell where a
mistake was made and sometimes restarting is the best option. The
assembler will be so nice.