feat(atom): print line numbers
I need a reference for line numbers so that I can navigate to lines. I
decided to use a pretty simple algorithm for printing base 10 numbers,
leaving in the leading zeros. In some ways, it is kind of ugly, but it
also means that everything is aligned which is good for readability.

It is good enough.