fix(atom): null terminate when reading lines
Without adding a null terminator, the new value can end up embedded in
the old if it was longer. If a node contained "AAAAAAAAAA" and then you
read in "HALT", you would get "HALTAAAAAA".

This was not a problem when inserting always allocated a new node (which
was zeroed out), but now that they are reused there can be garbage. This
should have always been done for safety.

This fix is small in assembly, but requires a lot of work to
re-assemble. Having an assembler would be so nice, even if it were slow.