docs(atom): add ideal assembly design
These are my thoughts on an ideal assembly language. I have marked this
file as ideal, and will be rebuilding it throughout the implantation
process. The real doc will be up to date with the implementation.

The assembly refers to a carry jump. This is technically just the same
conditions as the less than jump, but it is easier to understand if you
use two separate instructions. To that end, the less than entry was
duplicated in the jump types table so that the carry could be written
out explicitly.